rename hook file name
